configs: Add libapexsupport.so to sparse to keep appsupport happy.
[config] Add libapexsupport.so to sparse to keep appsupport happy. JB#63077

Due to version differences we need to copy the file to this location.
The files are identical copies of the /system variants.
Bind mounting etc is not an option because it would break updates.
We opted to not copy them during installation or appsupport startup
since then we would need to keep track of a list of files that have
been copied over and removing them would be annoying.